Add computing recommendation statistics
Application iterates over the tweets in the database, and for each
calculates number of users recommended in the tweet body. Creates
histogram of recommendation counts per tweet.

Important: It utilizes Hibernate ScrollableResults API, in order to
avoid loading the whole table to memory, which turns out to be
prohibitively expensive.