Add "can't get tweets" field to User
It turns out that 401 code while trying to fetch recent tweets of
certain users is usually due to the profile being protected. It makes
little sense to retry in such cases, and to prevent multiple requests
wasted for it for every application launch, special field to denote
this state is added.